requirements.txt
aniso8601==4.0.1
argon2-cffi==18.3.0
blinker==1.4
cffi==1.11.5
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.7
Flask-JWT-Extended==3.14.0
Flask-Mail==0.9.1
flask-marshmallow==0.9.0
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
marshmallow==2.16.3
pkg-resources==0.0.0
pycparser==2.19
PyJWT==1.7.1
pytz==2018.7
six==1.12.0
SQLAlchemy==1.2.15
Werkzeug==0.14.1
